+++ examples/general/accounts.php
@@ -1,0 +1,24 @@
+<?php
+
+use Mailtrap\Config;
+use Mailtrap\Helper\ResponseHelper;
+use Mailtrap\MailtrapClient;
+
+require __DIR__ . '/../vendor/autoload.php';
+
+// your API token from here https://mailtrap.io/api-tokens
+$apiKey = getenv('MAILTRAP_API_KEY');
+$mailtrap = new MailtrapClient(new Config($apiKey));
+
+/**
+ * Get a list of your Mailtrap accounts.
+ *
+ * GET https://mailtrap.io/api/accounts
+ */
+try {
+    $response = $mailtrap->general()->accounts()->getList();
+
+    var_dump(ResponseHelper::toArray($response)); // body (array)
+} catch (Exception $e) {
+    echo 'Caught exception: ',  $e->getMessage(), "\n";
+}