+++ src/Bridge/Transport/MailtrapTransportFactory.php
@@ -7,6 +7,7 @@ namespace Mailtrap\Bridge\Transport;
 use Mailtrap\Api\AbstractApi;
 use Mailtrap\Config;
 use Mailtrap\Exception\RuntimeException;
+use Mailtrap\MailtrapClient;
 use Mailtrap\MailtrapSandboxClient;
 use Mailtrap\MailtrapSendingClient;
 use Symfony\Component\HttpClient\Psr18Client;
@@ -31,17 +32,17 @@ final class MailtrapTransportFactory extends AbstractTransportFactory
             ->setHost('default' === $dsn->getHost() ? AbstractApi::SENDMAIL_HOST : $dsn->getHost())
             ->setHttpClient(null === $this->client ? null : new Psr18Client($this->client))
         ;
-        $mailTrapClient = stripos($config->getHost(), AbstractApi::SENDMAIL_HOST) !== false
-            ? new MailtrapSendingClient($config)
-            : new MailtrapSandboxClient($config);
+        $mailtrapClient = stripos($config->getHost(), AbstractApi::SENDMAIL_HOST) !== false
+            ? (new MailtrapClient($config))->sending()
+            : (new MailtrapClient($config))->sandbox();
 
-        if ($mailTrapClient instanceof MailtrapSandboxClient && null === $inboxId) {
+        if ($mailtrapClient instanceof MailtrapSandboxClient && null === $inboxId) {
             throw new RuntimeException(
                 'You cannot send email to the SanBox with empty "inboxId" param. Example -> "MAILER_DSN=mailtrap+api://APIKEY@sandbox.api.mailtrap.io?inboxId=1234"'
             );
         }
 
-        return new MailtrapApiTransport($mailTrapClient, $inboxId, $this->dispatcher, $this->logger);
+        return new MailtrapApiTransport($mailtrapClient, $inboxId, $this->dispatcher, $this->logger);
     }
 
     protected function getSupportedSchemes(): array