+++ src/MailtrapClientInterface.php
@@ -4,14 +4,7 @@ declare(strict_types=1);
 
 namespace Mailtrap;
 
-/**
- * @method  Api\Sandbox\SandboxEmails|Api\Sending\SendingEmails  emails
- */
 interface MailtrapClientInterface
 {
-    public function __construct(ConfigInterface $config);
-
-    public function getConfig(): ConfigInterface;
-
-    public function getApiClassByName(string $name): ?string;
+    public const API_MAPPING = [];
 }