+++ src/MailtrapSendingClient.php
@@ -7,22 +7,13 @@ namespace Mailtrap;
 use Mailtrap\Api;
 
 /**
- * @method  Api\Sending\SendingAccount  accounts
- * @method  Api\Sending\SendingEmails   emails
- * @method  Api\Sending\SendingUser     users
+ * @method  Api\Sending\Emails   emails
  *
- * Class MailtrapClient
+ * Class MailtrapSendingClient
  */
-class MailtrapSendingClient extends AbstractMailtrapClient
+final class MailtrapSendingClient extends AbstractMailtrapClient
 {
-    private const API_MAPPING = [
-        'accounts' => Api\Sending\SendingAccount::class,
-        'emails' => Api\Sending\SendingEmails::class,
-        'users' => Api\Sending\SendingUser::class,
+    public const API_MAPPING = [
+        'emails' => Api\Sending\Emails::class,
     ];
-
-    public function getApiClassByName(string $name): ?string
-    {
-        return !empty(self::API_MAPPING[$name]) ? self::API_MAPPING[$name] : null;
-    }
 }